backgroundimg animation added
BackgroundImg can now be pushed to the left and out of viewpoint with a delayed scrolling action.

NEXT: animate night img so that it effectlive pushes backgroundImg out of viewpoint and takes its place. Once that is done, I'll do the same with night img when it is replaced by the next img in the story.