sun zoom out feature achieved
So, I was able 2 update the animation so that it starts zoomed in to the sun completed and only after scrolling begins does the view zoom out and we see the entire 1st background.

NEXT: i'll c if it's possible to have more control over the animations so that the reverse debocles can b resolved for a buttery smooth animation that i've seen in other websites. Also, I wud get a new delivery img. After that, readd the mascot into the whole shabang.