partial fix to animation of upward/reverse scrolling
Ok, titile says all. I've updated the 1st gsap.timeline and the reverse scrolling also shows smooth animation in the backwards direction. the secret was the only the img on the x=0 horizontal position should have a z-index of 0. the img that goes off screen should have a z-index of -1

NEXT UP: I'll do that same for the other gsap.timeline as well to fix its reverse animation as well.