added extra imgs to aboutUs endppoint experiment
So, I was able to add two mroe imgs to the aboutUs endpoint experiment. Pretty cool that it's working out just fine!
NEXT UP: i'll try 2 give the horizontal slider effect the the imgs in the experiment. idea is basically 2 take this experiment as far as possible and combine this code with the root endpoint home page code.