horizontal scroll achieved
ok, so i was able to get the night img to horizontally scroll into the viewpoint instead of the usual vertical entry.

NEXT UP: i'll do the same thing for hte delivery img as well. Once that is done, i'll compare it with how I have it done currently in home endpoint and merge the two scripts together. Then, I can start playing with the last img which has more animations in it.