last img resolved in experiment
Ok, so i was able to finish the last img with the banner and everything.

NEXT: I have to turn the banner into a link which leads to the contact page. One the contact page, I'll have a sample contact form, and also a phony(pun-intended) number, along with some social media links.