link to contact page achieved!
So, i've now created a simple link from the banner to the /contact endpoint. Also, the contact page has been given a simple navbar.

NEXT UP: i'll readjust the size of the scroll story margins and other values in the aboutUs endpoint. Once that is done, then i will move that code to the root endpoint for hte homepage.