submitting the website worked! kinda
So, updating the endpoints in the app.js and navbar.js did allow me to publish my website! this is awesome!

not next up, but soon, I'll need to change scroll positions again for the about us endpoint with 'vh', or 'em', or something instead of pixels so that the animation works correctly for screens of all sizes. Also, the webpage needs to b made mobile friendly.

but b4 that,
NEXT UP: create an awesome contact form. Also, awesome! i do have a free source to upload frontend websites to now! All i need is to load jar files in lightsail when i decide it's time to start  creating full Stack websites :)